+++ packages/nadle/test/__fixtures__/npm-task/nadle.config.ts
@@ -1,0 +1,5 @@
+import { tasks, NpmTask } from "nadle";
+
+tasks.register("pass", NpmTask, { args: ["exec", "--", "tsc", "./src/pass.ts", "--noEmit", "--pretty"] });
+tasks.register("fail", NpmTask, { args: ["exec", "--", "tsc", "./src/fail.ts", "--noEmit", "--pretty"] });
+tasks.register("echo", NpmTask, { args: ["exec", "--", "echo", "hello"] });